Factor our ScrollSpy from Toc into separate component (#5520)
* Factor our ScrollSpy from Toc into separate component

* Factor our ScrollSpy from Toc into separate component

* Fix line height

* Fix line heights
