+++ src/site/_includes/partials/toc-side.njk
@@ -2,8 +2,10 @@
 {% if tocContents %}
   <nav class="course-toc toc flow scrollbar display-none xl:display-block" data-type="side" aria-label="{{ 'i18n.toc.on_this_page' | i18n(locale) }}">
     <div class="course-toc__heading font-google-sans weight-medium">{{ 'i18n.toc.on_this_page' | i18n(locale) }}</div>
-    <div class="toc__wrapper flow-recursive">
-      {{ tocContents | safe }}
-    </div>
+    <web-scroll-spy>
+      <div class="toc__wrapper flow-recursive">
+        {{ tocContents | safe }}
+      </div>
+    </web-scroll-spy>
   </nav>
 {% endif %}
