+++ src/styles/components/_course-toc.scss
@@ -52,6 +52,10 @@
   &__heading {
     font-size: px-to-rem(14px);
   }
+
+  .w-scroll-spy__active {
+    font-weight: 700;
+  }
 }
